requirements.txt
flask>=3.1.3
stripe>=8.0.0
python-dotenv>=1.0.0
